#9 - the names of some controllers and services have been changed, removed some unnecessary import, some comments were made
