+++ src/main/java/com/example/telegrambotinsurance/service/BotService.java
@@ -2,7 +2,6 @@ package com.example.telegrambotinsurance.service;
 
 import com.example.telegrambotinsurance.exception.BotNotFoundException;
 import com.example.telegrambotinsurance.modelbot.AbstractBot;
-import com.example.telegrambotinsurance.modelbot.Message;
 
 public interface BotService {
     //Метод ищет бота по токену
